+++ core/src/test/java/org/realityforge/arez/test/ExternalApiTest.java
@@ -60,6 +60,14 @@ public class ExternalApiTest
   }
 
   @Test
+  public void areRepositoryResultsUnmodifiable()
+  {
+    assertFalse( Arez.areRepositoryResultsModifiable() );
+    ArezTestUtil.makeRepositoryResultsModifiable();
+    assertTrue( Arez.areRepositoryResultsModifiable() );
+  }
+
+  @Test
   public void areSpiesEnabled()
   {
     assertTrue( Arez.areSpiesEnabled() );