+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithExplicitIdRepository.java
@@ -11,6 +11,7 @@ import java.util.stream.Stream;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.realityforge.arez.Arez;
 import org.realityforge.arez.Disposable;
 import org.realityforge.arez.Observable;
 import org.realityforge.arez.annotations.Action;
@@ -26,9 +27,6 @@ import org.realityforge.braincheck.Guards;
     singleton = true
 )
 public class RepositoryWithExplicitIdRepository implements RepositoryWithExplicitIdBaseRepositoryExtension {
-  private static final boolean $$arez$$_IMMUTABLE_RESULTS = "true".equals( System.getProperty( "arez.repositories_return_immutables", String.valueOf( System.getProperty( "arez.environment", "production" ).equals( "development" ) ) ) );
-  ;
-
   private final HashMap<Integer, RepositoryWithExplicitId> $$arez$$_entities = new HashMap<>();
   ;
 
@@ -118,7 +116,7 @@ public class RepositoryWithExplicitIdRepository implements RepositoryWithExplici
    */
   @Nonnull
   protected final List<RepositoryWithExplicitId> wrap(@Nonnull final List<RepositoryWithExplicitId> list) {
-    return $$arez$$_IMMUTABLE_RESULTS ? Collections.unmodifiableList( list ) : list;
+    return Arez.areRepositoryResultsModifiable() ? Collections.unmodifiableList( list ) : list;
   }
 
   /**