+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithMultipleCtorsRepository.java
@@ -11,6 +11,7 @@ import java.util.stream.Stream;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.realityforge.arez.Arez;
 import org.realityforge.arez.Disposable;
 import org.realityforge.arez.Observable;
 import org.realityforge.arez.annotations.Action;
@@ -25,9 +26,6 @@ import org.realityforge.braincheck.Guards;
     singleton = true
 )
 public class RepositoryWithMultipleCtorsRepository implements RepositoryWithMultipleCtorsBaseRepositoryExtension {
-  private static final boolean $$arez$$_IMMUTABLE_RESULTS = "true".equals( System.getProperty( "arez.repositories_return_immutables", String.valueOf( System.getProperty( "arez.environment", "production" ).equals( "development" ) ) ) );
-  ;
-
   private final HashMap<Long, RepositoryWithMultipleCtors> $$arez$$_entities = new HashMap<>();
   ;
 
@@ -126,7 +124,7 @@ public class RepositoryWithMultipleCtorsRepository implements RepositoryWithMult
    */
   @Nonnull
   protected final List<RepositoryWithMultipleCtors> wrap(@Nonnull final List<RepositoryWithMultipleCtors> list) {
-    return $$arez$$_IMMUTABLE_RESULTS ? Collections.unmodifiableList( list ) : list;
+    return Arez.areRepositoryResultsModifiable() ? Collections.unmodifiableList( list ) : list;
   }
 
   /**